Move the display catalog out of the island layout model
AgentIslandDisplayCatalog observes screen-parameter notifications and
owns the live NSScreen mapping, so it belongs with the controller in
BusinessLogic rather than beside the pure geometry in Models. Drop the
unused AgentIslandNotchLayout.rootWidth.
