+++ supacode/Features/ActiveAgents/Models/AgentIslandScreen.swift
@@ -1,6 +1,4 @@
-import AppKit
 import CoreGraphics
-import Observation
 
 struct AgentIslandScreenDescriptor: Equatable, Identifiable {
   let id: String
@@ -38,10 +36,6 @@ struct AgentIslandNotchLayout: Equatable {
     self.compactWidth = compactWidth
     compactHeight = max(Self.minimumCompactHeight, cutoutSize.height)
     wingWidth = (compactWidth - cutoutSize.width) / 2
-  }
-
-  var rootWidth: CGFloat {
-    max(420, compactWidth)
   }
 }
 
@@ -109,63 +103,6 @@ enum AgentIslandScreenLayout {
       y: top - contentSize.height,
       width: contentSize.width,
       height: contentSize.height
-    )
-  }
-}
-
-@MainActor
-@Observable
-final class AgentIslandDisplayCatalog {
-  static let shared = AgentIslandDisplayCatalog()
-
-  private(set) var screens: [AgentIslandScreenDescriptor] = []
-  private var observers: [NSObjectProtocol] = []
-
-  init(notificationCenter: NotificationCenter = .default) {
-    refresh()
-    let observer = notificationCenter.addObserver(
-      forName: NSApplication.didChangeScreenParametersNotification,
-      object: nil,
-      queue: .main
-    ) { [weak self] _ in
-      MainActor.assumeIsolated {
-        self?.refresh()
-      }
-    }
-    observers.append(observer)
-  }
-
-  func refresh() {
-    screens = NSScreen.screens.compactMap(Self.descriptor(for:))
-  }
-
-  func descriptor(for screen: NSScreen?) -> AgentIslandScreenDescriptor? {
-    guard let screen else { return nil }
-    return Self.descriptor(for: screen)
-  }
-
-  private static func descriptor(for screen: NSScreen) -> AgentIslandScreenDescriptor? {
-    guard
-      let number = screen.deviceDescription[NSDeviceDescriptionKey("NSScreenNumber")] as? NSNumber
-    else { return nil }
-    let displayID = CGDirectDisplayID(number.uint32Value)
-    guard let displayUUID = CGDisplayCreateUUIDFromDisplayID(displayID)?.takeRetainedValue() else {
-      return nil
-    }
-    let id = CFUUIDCreateString(nil, displayUUID) as String
-    let notchFrame = AgentIslandScreenLayout.notchFrame(
-      screenFrame: screen.frame,
-      safeAreaTopInset: screen.safeAreaInsets.top,
-      auxiliaryTopLeftArea: screen.auxiliaryTopLeftArea,
-      auxiliaryTopRightArea: screen.auxiliaryTopRightArea
-    )
-    return AgentIslandScreenDescriptor(
-      id: id,
-      name: screen.localizedName,
-      frame: screen.frame,
-      visibleFrame: screen.visibleFrame,
-      isBuiltIn: CGDisplayIsBuiltin(displayID) != 0,
-      notchFrame: notchFrame
     )
   }
 }
